Re-arranged directory structure, and created a separate structure for email related files
